feat: introduce API for RetryStrategy
This sets up the wiring for RetryStrategy in the Configuration API.
It doesn't actually do the final plumbing to connect the configuration
to the retry interceptor because we have a known bug in the interceptor
that I want to get fixed before we add the complexity.  But this at
least gets the API into what I hope will be the final state.