WIP: working proof-of-concept for middleware.
The goal of this commit is to illustrate a workable path forward
for adding the Middleware API, and make sure that it can be done
without any backward-incompatible changes.  I feel comfortable that
it can.

This POC sketches one possible API we could use, which I am not
content with.  The API included in this commit largely mimics
the Interceptor life cycle, where individual interceptors / middlewares
would be given an opportunity to respond immediately at different
phases of the lifecycle: when metadata becomes available, when the
request/response body becomes available, when the status becomes
available.  Each of these hooks provides a mechanism for the
interceptor/middleware to modify the single relevant state object
for that point in the life cycle (i.e. either metadata, request/response,
or state), but they are called asynchronously and the order is
not entirely predictable if there are async functions involved
in the interceptor / middleware implementation.

The motivation for trying out this API is beecause it allows the
interceptors/middlewares to modify the request/response as it
flows through.

After having played with this for a while I no longer believe that
the ability to modify those is worth the additional UX complexity
that it imposes for a Middleware API.

Instead, we should tweak this POC Middleware API to reduce
the number of functions on the interface.  The MiddlewareRequestHandler
should get invoked once for the request (at the point when both
the metadata and the request object are available) and then once more
for the response (at the point when all three of the response status,
metadata, and response object are available).  We don't need to give
the Middlewares the ability to modify these; they can be read-only,
which will allow us to change the return type to void and will simplify
a lot of the chaining that exists in this commit.

I feel confident that the improved API described above can be achieved
in a backward incompatible way, so I think it's okay to shelve this
for now.  I'm especially interested to see how the interceptors might
change when we add support for streaming (as opposed to unary) APIs.
This will happen when we add pubsub support, so that might be a good
time to revisit this.